+++ kernel/helper/info_text/info_text.h
@@ -17,4 +17,4 @@ enum class InfoTextType {
 void printInfoLine(enum InfoTextType textType, const char* infoText);
 void print_separator();
 
-#endif
+#endif // INFO_HELPER_H