+++ kernel/io/timer/timer.h
@@ -12,4 +12,4 @@ uint64_t get_ticks_in_ms();
 void sleep_sec(const uint64_t sec);
 void sleep_ms(const uint64_t ms);
 
-#endif
+#endif // TIMER_H