+++ kernel/shell/args.h
@@ -9,4 +9,4 @@ const char* getCurrTokenFromBuffer(const int inputBufferCursorAt_X, const uint64
 const char* getInputArgs(const uint64_t inputLength, char inputBuffer[TERMINAL_BUFFER_SIZE]);
 bool executeCommand(char inputBuffer[TERMINAL_BUFFER_SIZE], const char* args);
 
-#endif
+#endif // ARGS_H